requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.3
gunicorn==20.1.0
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.2.0
pip
Pillow
pytesseract
pdf2image
